Make custom bot class in seprate file
